+++ example/lib/test.dart
@@ -1,4 +1,10 @@
 class Test {
-  const Test({this.name});
+  const Test({this.name, this.nestedTest});
   final String name;
+  final NestedTest nestedTest;
+}
+
+class NestedTest {
+  const NestedTest({this.ze});
+  final String ze;
 }